Add support for Nested Content
When there was a Vorto property inside a Nested Content one, Vorto
couldn't read the property guid because that information is inside the
doctype that the Nested Content property has configured.

To fix that, now when there's a document type that has
Vorto + Nested Content, Vorto also looks in all the properties of every
doc type of every nested content. It has to be every doc type because
one nested content can have multiple doc types configured.