Allow to get Vorto value by language
Sometimes the callers know the language but not the culture of the
Vorto value they want, e.g., the current thread culture is en-US, but
the Umbraco culture is en-GB, so when they try to read the English value
they will get an empty string.

This allows the caller to pass only a language, i.e.,
Model.Content.GetVortoValue("property", "en") and the extension method
takes care to get the correct value.