Add support for Nested Content
When there was a Vorto property inside a Nested Content one, Vorto
couldn't read the property guid because that information is inside the
doctype that the Nested Content property has configured.

To fix that, now when Vorto doesn't found the property in the current
document type, it looks for it in all the document types.