Improve property look up performace
When a Nested Content with a Vorto property has a lot of items, looping
through the properties of all the document types could be an expensive
operation, so Vorto saves all the properties it's already read in memory
cache to avoid doing that.

The cache expires when it hasn't been used in 30 minutes.